requirements.txt
qdrant-client==1.9.0
azure-identity==1.16.0
aiohttp==3.9.5
fastapi==0.110.2
uvicorn[standard]==0.29.0